allow lua to change line number styles
Export the new function win:style_lineno(style, lineno) to lua.
Line number styles are stored in a sorted doubly linked list in the
corresponding UiTermWin struct.
Setting a line number's style to win.STYLE_DEFAULT resets the
style and frees the style descriptor.

This is useful to present information to the user without
cluttering or changing the interface to much.
initial commit adding styles to line numbers