lua: add functions to switch between byte offset and line number/column
In vis there are two representations for positions in the text:

* the 0-based byte index
* a tuple of 1-based line number and 1-based column

Export the text_ functions converting between both representations to
Lua.

Lua currently has no good way to convert between the representation.
Setting the selection based on one and retrieving the other possibly
changes the viewport.
Which can result in view jumps noticeable by the user or is straight
up not possible for example during WIN_HIGHLIGHT events.

The second variant of iterating all lines is rather expensive because
each line has to be pushed to Lua as string.